--- qp_klp/tests/sample_info_file_1_blanks.tsv
@@ -1,13 +1,0 @@
-sample_name	collection_timestamp	elevation	empo_1	empo_2	empo_3	env_biome	env_feature	env_material	env_package	geo_loc_name	host_subject_id	latitude	longitude	sample_type	scientific_name	taxon_id	description	title	dna_extracted	physical_specimen_location	physical_specimen_remaining
-BLANK.None.20C.BLANK.1	2022-1-19	193	Control	Negative	Sterile water blank	urban biome	research facility	sterile water	misc environment	USA:CA:San Diego	BLANK.None.20C.BLANK.1	32.5	-117.25	control blank	metagenome	256318	BLANK.None.20C.BLANK.1	PLUS_Urobiome_Validation_Vaginal	TRUE	UCSD	FALSE
-BLANK.None.20C.BLANK.2	2022-1-19	193	Control	Negative	Sterile water blank	urban biome	research facility	sterile water	misc environment	USA:CA:San Diego	BLANK.None.20C.BLANK.2	32.5	-117.25	control blank	metagenome	256318	BLANK.None.20C.BLANK.2	PLUS_Urobiome_Validation_Vaginal	TRUE	UCSD	FALSE
-BLANK.AssayAssure.20C.BLANK.1	2022-1-19	193	Control	Negative	Sterile water blank	urban biome	research facility	sterile water	misc environment	USA:CA:San Diego	BLANK.AssayAssure.20C.BLANK.1	32.5	-117.25	control blank	metagenome	256318	BLANK.AssayAssure.20C.BLANK.1	PLUS_Urobiome_Validation_Vaginal	TRUE	UCSD	FALSE
-BLANK.AssayAssure.20C.BLANK.2	2022-1-19	193	Control	Negative	Sterile water blank	urban biome	research facility	sterile water	misc environment	USA:CA:San Diego	BLANK.AssayAssure.20C.BLANK.2	32.5	-117.25	control blank	metagenome	256318	BLANK.AssayAssure.20C.BLANK.2	PLUS_Urobiome_Validation_Vaginal	TRUE	UCSD	FALSE
-BLANK.EtOH.20C.BLANK.1	2022-1-19	193	Control	Negative	Sterile water blank	urban biome	research facility	sterile water	misc environment	USA:CA:San Diego	BLANK.EtOH.20C.BLANK.1	32.5	-117.25	control blank	metagenome	256318	BLANK.EtOH.20C.BLANK.1	PLUS_Urobiome_Validation_Vaginal	TRUE	UCSD	FALSE
-BLANK.EtOH.20C.BLANK.2	2022-1-19	193	Control	Negative	Sterile water blank	urban biome	research facility	sterile water	misc environment	USA:CA:San Diego	BLANK.EtOH.20C.BLANK.2	32.5	-117.25	control blank	metagenome	256318	BLANK.EtOH.20C.BLANK.2	PLUS_Urobiome_Validation_Vaginal	TRUE	UCSD	FALSE
-BLANK5.12C	2022-1-19	193	Control	Negative	Sterile water blank	urban biome	research facility	sterile water	misc environment	USA:CA:San Diego	BLANK5.12C	32.5	-117.25	control blank	metagenome	256318	BLANK5.12C	PLUS_Urobiome_Validation_Vaginal	TRUE	UCSD	FALSE
-BLANK5.12D	2022-1-19	193	Control	Negative	Sterile water blank	urban biome	research facility	sterile water	misc environment	USA:CA:San Diego	BLANK5.12D	32.5	-117.25	control blank	metagenome	256318	BLANK5.12D	PLUS_Urobiome_Validation_Vaginal	TRUE	UCSD	FALSE
-BLANK5.12E	2022-1-19	193	Control	Negative	Sterile water blank	urban biome	research facility	sterile water	misc environment	USA:CA:San Diego	BLANK5.12E	32.5	-117.25	control blank	metagenome	256318	BLANK5.12E	PLUS_Urobiome_Validation_Vaginal	TRUE	UCSD	FALSE
-BLANK5.12F	2022-1-19	193	Control	Negative	Sterile water blank	urban biome	research facility	sterile water	misc environment	USA:CA:San Diego	BLANK5.12F	32.5	-117.25	control blank	metagenome	256318	BLANK5.12F	PLUS_Urobiome_Validation_Vaginal	TRUE	UCSD	FALSE
-BLANK5.12G	2022-1-19	193	Control	Negative	Sterile water blank	urban biome	research facility	sterile water	misc environment	USA:CA:San Diego	BLANK5.12G	32.5	-117.25	control blank	metagenome	256318	BLANK5.12G	PLUS_Urobiome_Validation_Vaginal	TRUE	UCSD	FALSE
-BLANK5.12H	2022-1-19	193	Control	Negative	Sterile water blank	urban biome	research facility	sterile water	misc environment	USA:CA:San Diego	BLANK5.12H	32.5	-117.25	control blank	metagenome	256318	BLANK5.12H	PLUS_Urobiome_Validation_Vaginal	TRUE	UCSD	FALSE
