+++ qp_klp/tests/test_step.py
@@ -1,0 +1,15 @@
+# -----------------------------------------------------------------------------
+# Copyright (c) 2014--, The Qiita Development Team.
+#
+# Distributed under the terms of the BSD 3-clause License.
+#
+# The full license is in the file LICENSE, distributed with this software.
+# -----------------------------------------------------------------------------
+from unittest import TestCase
+from qp_klp.Step import Step
+
+
+class BaseStepTests(TestCase):
+    def test_creation(self):
+        step = Step(None, None, None, None)
+        self.assertFalse(step is None)
